Move first argument in functional
Doesn't add any dependencies
