+++ libs/functional/include_compatibility/hpx/util/first_argument.hpp
@@ -1,0 +1,21 @@
+//  Copyright (c) 2019 Ste||ar Group
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/config.hpp>
+#include <hpx/functional/config/defines.hpp>
+#include <hpx/functional/first_argument.hpp>
+
+#if defined(HPX_FUNCTIONAL_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_MSVC)
+#pragma message( \
+    "The header hpx/util/first_argument.hpp is deprecated, \
+    please include hpx/functional/first_argument.hpp instead")
+#else
+#warning \
+    "The header hpx/util/first_argument.hpp is deprecated, \
+    please include hpx/functional/first_argument.hpp instead"
+#endif
+#endif
